requirements.txt
flask
line-bot-sdk
langdetect
googletrans==4.0.0-rc1
